chore(gameplay): clean up for more intuitive use
