+++ src/ws/commands/game.start.ts
@@ -5,7 +5,8 @@ export default function ({
     room: { game },
   },
 }: WebSocketManager) {
-  setTimeout(() => {
-    game?.me?.start();
-  }, game?.me?.player.pregameTime);
+  if (game?.me)
+    setTimeout(() => {
+      game?.me?.start();
+    }, game.me.player.options.prestart + /*this.isNew*/ (true ? game.me.player.options.precountdown : 0) + (game.me.player.options.countdown ? game.me.player.options.countdown_interval * game.me.player.options.countdown_count : 0));
 }
